+++ lecture/ch4/ch4/ch4.vcxproj.filters
@@ -21,5 +21,8 @@
     <ClCompile Include="1_A_combination.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="1_A_per_rec.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
